Make new structure for the CLI. Implement module add subcommand
