+++ src/cli/mod.rs
@@ -1,0 +1,2 @@
+pub mod main_command;
+pub mod module_subcommand;
